Move pull list code to a seperate file
